PID
* Created NaivePIDController class (which does PID)
* Put Encoders in the Chassis class
* Moved control code from the Chassis class to Commands
* Modified MoveStraight command to use PID
* In the Chassis subsystems, created PID controllers for both position
and velocity of each gearbox.
--
Problems:
* When naming the pidcontroller P/I/D values in smartdashboard (in the
Project... method in the naivepidcontroller class), I expect the fields
to show up as "[name] P", "[name] I", and "[name]" D]. However, the
names show up as simply "P", "I", and "D"--without the name.
* SmartDashboard only supports numbers up to 3 places past the decimal.
Scale the PID values so they fit.
* It doesn't actually work. Haven't had a chance to look into it yet
(what the actual problems are). It would be nice to reimage the robot to
enable NetConsole, so we can printf output.