Finished basis of Intake Mechanism Subsystem
Finished basic filler code for intake mechanism subsystem and modified
files accordingly. Right now there is a command that will (hopefully)
extend and retract the arms as a toggle and will start and stop the
motor for grabbing the ball. Code for the actual execution is
rudimentary and probably should be replaced/refined later. Will end work
on this and probably go back to catapult test for a while.