Manual override for StopperWinch
ATTENTION: now the HIGH and LOW stop positions are set in degrees in the
SmartDashboard. The HIGH stopper position can be "recalibrated" using
the gamepad's thimbstick if needed. The LOW position will be relative to
the HIGH.

Also the Preferences are not used anymore. All presets are defined
("hardcoded") as const in the Robot class.